+++ components/function-controller/internal/controllers/kubernetes/remote_registry_cfg_webhook_test.go
@@ -1,4 +1,4 @@
-package serverless
+package kubernetes
 
 import (
 	"context"
